+++ lib/data/data_sources/cashu_wallet_data_source.dart
@@ -3,13 +3,15 @@ import 'dart:io';
 import 'package:cdk_flutter/cdk_flutter.dart' as cdk;
 import 'package:path_provider/path_provider.dart';
 
+import '../services/wallet/cdk_init_service.dart';
+
 class CashuWalletDataSource {
   final cdk.WalletRepository _wallet;
 
   CashuWalletDataSource._(this._wallet);
 
   static Future<CashuWalletDataSource> init() async {
-    await cdk.CdkFlutter.init();
+    await CdkInitService.ensureInitialized();
     final path = await getApplicationDocumentsDirectory();
     final mnemonicFile = File('${path.path}/mnemonic.txt');
 
@@ -21,8 +23,8 @@ class CashuWalletDataSource {
       await mnemonicFile.writeAsString(mnemonic);
     }
 
-    final db =
-        await cdk.WalletDatabase.newInstance(path: '${path.path}/wallet.sqlite');
+    final db = await cdk.WalletDatabase.newInstance(
+        path: '${path.path}/wallet.sqlite');
     final wallet = await cdk.WalletRepository.newInstance(
       unit: 'sat',
       mnemonic: mnemonic,